requirements.txt
fastapi
uvicorn
sqlmodel
asyncpg
openai-whisper
torch
openai
python-jose[cryptography]
passlib[bcrypt]
websockets
python-dotenv
pydub
pydantic-settings
psycopg2-binary
aiofiles
python-multipart
ffmpeg-python
numpy
scipy
requests
PyJWT==2.8.0
deepgram-sdk